requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
redis>=5.0.3
PyYAML>=6.0
pydantic>=2.8
orjson>=3.10
tqdm>=4.66
python-multipart>=0.0.9

# Vector DB
chromadb>=0.5.3

# Embeddings (Pi/CPU friendly)
sentence-transformers>=3.1.0
transformers>=4.43
numpy>=1.26
scipy>=1.11

# Optional helpers
python-dotenv>=1.0
loguru>=0.7
faiss-cpu

